# src/main/resources/api.openapi.yml
openapi: 3.0.1
info:
  title: WePlay API
  version: '1.0'
...
paths:
...
  /sports-facilities:
    get:
      tags:
        - Sports Facilities
      summary: >
        Gets the list of sports facilities
      operationId: getSportsFacilities
      parameters:
        - name: filter_by_owner_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
        - name: total_sports_field.gt
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: total_sports_field.lt
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: >
            The list of sports facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SportsFacility'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /sports-facilities/{sportsFacilityId}:
    get:
      tags:
        - Sports Facilities
      summary: >
        Gets a sports facility by ID
      operationId: getSportsFacilityById
      parameters:
        - name: sportsFacilityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: >
            The sports facility identified by the ID provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsFacilityWithSportsFields'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /sports-facilities/{sportsFacilityId}/sports-fields:
    post:
      tags:
        - Sports Facilities
      summary: >
        Creates a sports field for the sports 
        facility identified by the ID provided
      operationId: createSportsField
      parameters:
        - name: sportsFacilityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        $ref: '#/components/requestBodies/SportsFieldRequestBody'
      responses:
        '201':
          description: >
            The created sports field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsField'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
  /sports-facilities/{sportsFacilityId}/reservations-summaries:
    get:
      tags:
        - Sports Facilities
      summary: >
        Generate a custom reservations summary for the sports facility identified by the ID provided
      operationId: generateReservationsSummary
      parameters:
        - name: sportsFacilityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
        - name: start_date
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/Date'
        - name: end_date
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Date'
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum:
              - total_reservations.asc
              - total_reservations.desc
              - accepted_reservations.asc
              - accepted_reservations.desc
              - rejected_reservations.asc
              - rejected_reservations.desc
              - total_revenue.asc
              - total_revenue.desc
              - sport
            default: total_revenue.desc
      responses:
        '201':
          description: >
            The reservation summary generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationsSummary'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
...
components:
  schemas:
    ID:
      type: integer
      format: int64
      readOnly: true
      minimum: 0
    Date:
      type: string
      format: date-time
...
    SportEnum:
      type: string
      enum: ['soccer', 'basket', 'volleyball', 'tennis']
    BaseSportsField:
      type: object
      required:
        - id
        - sportsFacilityId
        - name
        - sport
        - isIndoor
        - priceList
      properties:
        id:
          $ref: '#/components/schemas/ID'
        sportsFacilityId:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        sport:
          type: string
        isIndoor:
          type: boolean
        priceList:
          $ref: '#/components/schemas/SportsFieldPriceList'
    SoccerField:
      allOf:
        - $ref: '#/components/schemas/BaseSportsField'
        - type: object
          required:
            - soccerFieldType
          properties:
            soccerFieldType:
              type: string
              enum: ['5 a side', '8 a side', '11 a side']
    VolleyballField:
      allOf:
        - $ref: '#/components/schemas/BaseSportsField'
        - type: object
    BasketballField:
      allOf:
        - $ref: '#/components/schemas/BaseSportsField'
        - type: object
    TennisField:
      allOf:
        - $ref: '#/components/schemas/BaseSportsField'
        - type: object
          properties:
            tennisFieldType:
              type: string
              enum: ['red clay', 'grass', 'cement']
    SportsField:
      oneOf:
        - $ref: '#/components/schemas/BasketballField'
        - $ref: '#/components/schemas/VolleyballField'
        - $ref: '#/components/schemas/SoccerField'
        - $ref: '#/components/schemas/TennisField'
      discriminator:
        propertyName: sport
        mapping:
          soccer:     '#/components/schemas/SoccerField'
          basket:     '#/components/schemas/BasketballField'
          volleyball: '#/components/schemas/VolleyballField'
          tennis:     '#/components/schemas/TennisField'
    SportsFieldPriceList:
      type: object
      properties:
        pricePerHour:
          type: number
          format: float
          minimum: 0
    SportsFacility:
      type: object
      required:
        - name
        - id
        - ownerId
      properties:
        id:
          $ref: '#/components/schemas/ID'
        ownerId:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        totalSportsFields:
          type: integer
        address:
          $ref: '#/components/schemas/Address'
        phone:
          type: string
    SportsFacilityWithSportsFields:
      allOf:
        - $ref: '#/components/schemas/SportsFacility'
        - type: object
          properties:
            sportsFields:
              type: array
              items:
                $ref: '#/components/schemas/SportsField'
...
    Address:
      type: object
      required:
        - state
        - city
        - streetName
        - streetNumber
        - postcode
      properties:
        state:
          type: string
        city:
          type: string
        streetName:
          type: string
        streetNumber:
          type: string
        postcode:
          type: string
...
    ReservationsSummary:
      type: object
      required:
        - sportsFacilityId
        - sportsFacilityName
        - description
        - createdAt
        - sportsReservationReports
      properties:
        sportsFacilityId:
          $ref: '#/components/schemas/ID'
        sportsFacilityName:
          type: string
        description:
          type: string
        createdAt:
          $ref: '#/components/schemas/Date'
        sportsReservationReports:
          type: array
          items:
            $ref: '#/components/schemas/SportsReservationReport'
...
    SportsReservationReport:
      type: object
      required:
        - totalReservations
        - sport
        - totalRevenue
        - rejectedReservations
        - acceptedReservations
        - pendingReservations
      properties:
        totalReservations:
          type: integer
          format: int64
        sport:
          $ref: '#/components/schemas/SportEnum'
        totalRevenue:
          type: number
          format: float
        rejectedReservations:
          type: integer
          format: int64
        acceptedReservations:
          type: integer
          format: int64
        pendingReservations:
          type: integer
          format: int64
...
    Error:
      required:
        - detail
        - status
        - title
      properties:
        detail:
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
          type: string
        status:
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          example: 404
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized);
          example: Service Unavailable
          type: string
      type: object
...
  responses:
    400BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    404NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    SportsFieldRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SportsField'
security:
  - bearerAuth: []